requirements.txt
pandas
matplotlib
seaborn
re
underthesea
torch
transformer
streamlit